+++ lib/roundable.rb
@@ -7,6 +7,8 @@ module Roundable
   # greater than around 20 are usually quoted to three decimal places and
   # exchange rates greater than 80 are quoted to two decimal places.
   # Currencies over 5000 are usually quoted with no decimal places.
+  #
+  # https://en.wikipedia.org/wiki/Exchange_rate#Quotations
   def round(value)
     if value > 5000
       value.round
@@ -16,8 +18,13 @@ module Roundable
       Float(format('%<value>.3f', value: value))
     elsif value > 1
       Float(format('%<value>.4f', value: value))
-    else
+    # I had originally opted to round smaller numbers simply to five decimal
+    # places but introduced this refinement to handle an edge case where a
+    # lower-rate base currency like IDR produces less precise quotes.
+    elsif value > 0.0001
       Float(format('%<value>.5f', value: value))
+    else
+      Float(format('%<value>.6f', value: value))
     end
   end
 end